requirements.txt
autopep8==1.5.7
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
Flask==2.0.1
Flask-Cors==3.0.10
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
pypiwin32==223
python-vlc==3.0.12118
pyttsx3==2.90
pywin32==301
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
